[move] application/usecase → domain/application/usecase
