--- project/bodquest_2023/lib/application/usecase/firebase_options.dart
@@ -1,80 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      case TargetPlatform.windows:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for windows - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      default:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyBwj9hRzcZksLKIgwDQKWvhFxQ2uIhsrco',
-    appId: '1:663590594175:web:158f21ee8f5133a2d8814a',
-    messagingSenderId: '663590594175',
-    projectId: 'bod-quest-2023-11-17',
-    authDomain: 'bod-quest-2023-11-17.firebaseapp.com',
-    storageBucket: 'bod-quest-2023-11-17.appspot.com',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyDMCBLz3Q0msK0RhN4S-t_p37rsviGduro',
-    appId: '1:663590594175:android:46f348df0ef6236ad8814a',
-    messagingSenderId: '663590594175',
-    projectId: 'bod-quest-2023-11-17',
-    storageBucket: 'bod-quest-2023-11-17.appspot.com',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyAT2ByMwwun4tRj3tbAkl2DTpfHyvGS4Iw',
-    appId: '1:663590594175:ios:2eb93efb873b86e9d8814a',
-    messagingSenderId: '663590594175',
-    projectId: 'bod-quest-2023-11-17',
-    storageBucket: 'bod-quest-2023-11-17.appspot.com',
-    iosBundleId: 'com.example.bodquest2023',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyAT2ByMwwun4tRj3tbAkl2DTpfHyvGS4Iw',
-    appId: '1:663590594175:ios:beebc8e90e2ee7b5d8814a',
-    messagingSenderId: '663590594175',
-    projectId: 'bod-quest-2023-11-17',
-    storageBucket: 'bod-quest-2023-11-17.appspot.com',
-    iosBundleId: 'com.example.bodquest2023.RunnerTests',
-  );
-}
