+++ project/bodquest_2023/lib/domain/application/usecase/get_trainings_usecase_impl.dart
@@ -1,6 +1,6 @@
-import '../../domain/entity/training.dart';
-import '../../domain/repository/training_repository.dart';
-import '../../domain/usecase/get_trainings_usecase.dart';
+import '../../entity/training.dart';
+import '../../repository/training_repository.dart';
+import '../../usecase/get_trainings_usecase.dart';
 
 class GetTrainingsUsecaseImpl implements IGetTrainingsUsecase {
   final ITrainingRepository weightRepository;
