+++ project/bodquest_2023/lib/domain/application/usecase/get_weights_usecase_impl.dart
@@ -1,6 +1,6 @@
-import '../../domain/entity/weight.dart';
-import '../../domain/repository/weight_repository.dart';
-import '../../domain/usecase/get_weights_usecase.dart';
+import '../../entity/weight.dart';
+import '../../repository/weight_repository.dart';
+import '../../usecase/get_weights_usecase.dart';
 
 class GetWeightsUsecaseImpl implements IGetWeightsUsecase {
   final IWeightRepository weightRepository;
